requirements.txt
blinker==1.7.0
bokeh==3.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
docopt==0.6.2
Flask==3.0.0
Flask-Login==0.6.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipreqs==0.4.13
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tornado==6.4
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
xyzservices==2023.10.1
yarg==0.1.9
